requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
arrow==1.4.0
asgiref==3.10.0
asyncssh==2.22.0
atpublic==7.0.0
attrs==26.1.0
beautifulsoup4==4.13.4
bidict==0.23.1
billiard==4.2.4
binaryornot==0.4.4
blinker==1.9.0
blis==1.3.0
brotli==1.2.0
cachetools==7.0.5
catalogue==2.0.10
celery==5.6.2
certifi==2025.4.26
cffi==2.0.0
chardet==7.0.0
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
ConfigArgParse==1.7.1
configobj==5.0.9
contourpy==1.3.3
cookiecutter==2.7.0
cryptography==46.0.5
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
Django==5.2.7
docstring_parser==0.17.0
dpath==2.2.0
dulwich==1.1.0
dvc==3.67.0
dvc-data==3.18.3
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoints==0.4
faiss-cpu==1.13.2
fastapi==0.121.3
filelock==3.25.2
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
flatten-dict==0.4.2
flufl.lock==9.0.0
fonttools==4.59.1
fpdf==1.7.2
frozenlist==1.8.0
fsspec==2026.2.0
funcy==2.0
gevent==25.9.1
geventhttpclient==2.3.5
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grandalf==0.8
greenlet==3.2.4
grpcio==1.78.0
grpcio-status==1.71.2
gto==1.9.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.1
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.6.2
langchain==1.2.12
langchain-anthropic==1.4.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.20
langchain-google-genai==4.2.1
langchain-huggingface==1.2.1
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langcodes==3.5.0
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.20
language_data==1.3.0
locust==2.42.5
locust-cloud==1.29.4
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.5
mdurl==0.1.2
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==2.18.0
networkx==3.6.1
numpy==2.2.5
nvidia-ml-py==13.595.45
omegaconf==2.3.0
openai==2.29.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.3
pathspec==1.0.4
pillow==10.4.0
platformdirs==4.5.0
pluggy==1.6.0
preshed==3.0.10
proglog==0.1.11
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.1.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.13.1
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
pygit2==1.19.1
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.3
pypdf==6.9.2
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.3
python-slugify==8.0.4
python-socketio==5.15.0
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.30.0
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
scmrepo==3.6.1
semver==3.0.4
sentence-transformers==5.3.0
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
simple-websocket==1.1.0
six==1.17.0
smart_open==7.3.1
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
sqlparse==0.5.3
sqltrie==0.11.2
srsly==2.5.1
starlette==0.50.0
streamlit==1.55.0
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.14.0
torch==2.10.0
tornado==6.5.5
tqdm==4.67.1
transformers==5.3.0
typer==0.19.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uuid_utils==0.14.1
uv==0.9.3
uvicorn==0.38.0
vine==5.1.0
voluptuous==0.16.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.6.0
weasel==0.4.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.4
wrapt==1.17.3
wsproto==1.3.2
xxhash==3.6.0
yarl==1.23.0
youtube-transcript-api==1.2.4
zc.lockfile==4.0
zope.event==6.1
zope.interface==8.1.1
zstandard==0.25.0
